Release: v0.32.1
The `versionCode` has been bumped to `1000` so that users who have installed from F-Droid or GitHub should not have the app attempted to be updated by Google PlayStore and failing and also shown in PlayStore app updates list due to non-collaborative `v0.33` app release on PlayStore that set the `versionCode` higher than the latest F-Droid or GitHub `32` release. Unlike F-Droid, PlayStore does not check for difference in app APK signature before attempting to download and then failing to install due to signature mismatch.

- https://github.com/termux/termux-app/discussions/4000
- https://github.com/termux/termux-app/issues/4012
- https://github.com/termux-play-store/termux-apps/commit/b47c83d2d2277efd75b9c86ce59536771fdb65ee